port-serial: add blocked flag
Some ports have multiple roles, like Huawei USB sticks where
the QCDM port can be switched to PCM audio mode with AT^DDSETEX=2
when making voice calls.  We must ensure that during the voice call
no QCDM is used on the port, but we still want the modem to be able
to open the port object around when the voice call is done.